Move solvers into solvers/
