+++ pyjobshop/solvers/cpoptimizer/ConstraintsManager.py
