+++ pyjobshop/solvers/cpoptimizer/VariablesManager.py
