+++ pyjobshop/solvers/ortools/ObjectiveManager.py
