+++ pyjobshop/solvers/ortools/VariablesManager.py
